Improved template code

+++ cuda_addon/difi.cuh
@@ -11,7 +11,7 @@ void cudaCheckError(const char *message, const char *file, const int line);
 void copy_mesh(i3d::Model3D *model);
 
 void copy_distancemap(i3d::DistanceMap<double,i3d::cpu> *map);
-void transfer_distancemap(i3d::RigidBody *body, i3d::DistanceMap<double,i3d::cpu> *map);
+void transfer_distancemap(i3d::RigidBody *body, i3d::DistanceMap<float,i3d::cpu> *map);
 void dmap_test(i3d::RigidBody *body);
 void sphere_test(i3d::RigidBody *body, i3d::UniformGrid<i3d::Real,i3d::ElementCell,i3d::VertexTraits<i3d::Real>> &grid);
